+++ __tests__/manual/vb1.7.3/test-constants.ts
@@ -1,0 +1,2 @@
+export const TEST_VERSION = 'b1.7.3';
+export const TEST_MAPPING = 'feather';